ASoC: core: Add API to use DMI name as sound card name.
Intel DSP platform drivers are used by many different devices but are
difficult for userspace to differentiate. This patch adds an API to allow
the DMI name to be used as the sound card name, thereby helping userspace
load the correct UCM configuration.

This patch also allows for further differentiation as many devices that share
the same DMI name i.e. Minnowboards, UP boards may be configured with different
codecs or firmwares. The API supports flavoring the DMI name into the card
longname to provide the extra differentiation required for these devices.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>